serverr/requirements.txt
Flask
pymongo
dnspython
flask-cors
python-dotenv
bcrypt
pyjwt
